add common modal types and move the types there, use the config file for actual config
