+++ frontend/projects/upgrade/src/app/features/dashboard/feature-flags/modals/delete-feature-flag-modal/delete-feature-flag-modal.component.ts
@@ -9,13 +9,14 @@ import {
   MatDialogRef,
   MatDialogTitle,
 } from '@angular/material/dialog';
-import { CommonModalConfig } from '../../../../../shared-standalone-component-lib/components/common-modal/common-modal-config';
+
 import { FormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { FeatureFlagsService } from '../../../../../core/feature-flags/feature-flags.service';
 import { BehaviorSubject, Observable, Subscription, combineLatest, map } from 'rxjs';
 import { TranslateModule } from '@ngx-translate/core';
 import { CommonModule } from '@angular/common';
+import { CommonModalConfig } from '../../../../../shared-standalone-component-lib/components/common-modal/common-modal.types';
 
 @Component({
   selector: 'app-add-feature-flag-modal',
@@ -53,7 +54,7 @@ export class DeleteFeatureFlagModalComponent {
 
   constructor(
     @Inject(MAT_DIALOG_DATA)
-    public data: CommonModalConfig & { flagName: string; flagId: string },
+    public data: CommonModalConfig,
     public dialog: MatDialog,
     private featureFlagsService: FeatureFlagsService,
     public dialogRef: MatDialogRef<DeleteFeatureFlagModalComponent>
