+++ frontend/projects/upgrade/src/app/shared-standalone-component-lib/components/common-modal/common-modal-config.ts
@@ -1,20 +1,7 @@
-import { MatDialogConfig } from '@angular/material/dialog';
+import { environment } from '../../../../environments/environment';
 
-export interface CommonModalConfig<ParamsType = unknown> {
-  title: string;
-  cancelBtnLabel?: string;
-  primaryActionBtnLabel?: string;
-  primaryActionBtnColor?: string;
-  hideFooter?: boolean;
-  params?: ParamsType;
-}
-
-export interface CommonDialogMatDialogConfig extends MatDialogConfig {
-  data: CommonModalConfig;
-}
-
-export interface SimpleConfirmationModalParams {
-  message: string;
-  subMessage?: string;
-  subMessageClass?: 'info' | 'warn';
-}
+// see close-modal.interceptor.ts
+export const ENDPOINTS_TO_INTERCEPT_FOR_MODAL_CLOSE = [
+  environment.api.addFlagInclusionList,
+  environment.api.addFlagExclusionList,
+];
